move list to str function to utils
