+++ src/analyze/analyze_cutouts.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from omegaconf import DictConfig, OmegaConf
 
 # util imports
-from utils.pdf import PDFDrafter, add_grammar_and_capitlization_to_list
-from utils.utils import clear_directory, read_recipe, query_for_cutout_metadata
+from utils.pdf import PDFDrafter 
+from utils.utils import clear_directory, read_recipe, query_for_cutout_metadata, add_grammar_and_capitlization_to_list
 from utils.graphs import horizontal_bar_chart_plot, jitter_plot, vertical_bar_chart_plot, boolean_horizontal_bar_chart_plot
 
 log = logging.getLogger(__name__)
