+++ src/preprocess_cutouts.py
@@ -5,8 +5,7 @@ import numpy as np
 from tqdm import tqdm
 from omegaconf import DictConfig
 
-from utils.utils import index_cutouts_by_species
-from utils.pdf import add_grammar_and_capitlization_to_list
+from utils.utils import index_cutouts_by_species, add_grammar_and_capitlization_to_list
 
 log = logging.getLogger(__name__)
